seperated components into individual files
